Resolves conflicts by taking develop's versions throughout: main's
5 unique commits (older CI/CD, multi-tenant work) predate the same
features already landed and superseded on develop (idempotency keys,
BigDecimal money handling, gift card ledger, least-privilege DB role).
No main-only logic was lost - every conflicting file on main was a
strict subset of develop's version.

+++ src/main/java/com/finovago/p2p/dto/HoldResponse.java
@@ -1,5 +1,6 @@
 package com.finovago.p2p.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -16,5 +17,5 @@ public record HoldResponse(
     LocalDateTime expiresAt,
 
     @Schema(description = "Gift card balance still available for new holds after this operation. Populated on reserve only; null on capture/release.", example = "50.0")
-    Double remainingAvailableBalance
+    BigDecimal remainingAvailableBalance
 ) {}
